Contains several programs that answer different questions
Program 1 ask the user for the length and width of a rectangle and tells them its area. Program 2  takes a a Celsius temperature and converts it to Fahrenheit.  Program modifies an earlier version of a movie database program to ask the user about two different movies and offers a recommendation based on their rating.