Silver submission of Dataman
Bronze (max 85/100): 

Go through the existing program and test it.

(First, trade off with a classmate and try to use each others program. Next, show it to someone you know who doesn't write programs such as a friend or family member, and ask them to try it out. Are they able to navigate the menus easily? Do they get lost or stuck?)

Based on your tests, and feedback from users, go through and add "quality of life" improvements to the program. For example, all the menus should have a consistent look and feel. 

In addition, make sure that all code related to math is in the arithmetic source file, and all code related to temperature conversion is in the temperature source file.

Silver (max 95/100): 

Bronze, plus add Celcius -> Fahrenheit and F -> C functionality to the temperature converter.